further scan splitting
